✅ REDBLUE7 | Medium | Java | Runtime: N/A | Memory: N/A | 2026-09-16

+++ codechef/medium/REDBLUE7/solution.java
@@ -20,20 +20,20 @@ class Codechef
 	    }
 	    
 	    Arrays.sort(A);
-	    int Sr=0;
+	    long  Sr=0;
 	    int Sb= 0;
 	    int Cb =0;
 	    int Cr =0;
-	    int max =0 ;
+	    long max =0 ;
 	    
 	    
 	    for(int i =0 ; i< N ; i++){
 	        Sb += A[i];
-	      Cb = i + 1;
-           Sr = sum - Sb;
+	        Cb = i + 1;
+            Sr = sum - Sb;
             Cr = N - Cb;
 
-    int current = Sr * Cb + Sb * Cr;
+    long current = Sr * Cb + Sb * Cr;
 
     if (current > max) {
            max = current;
